day one of not being lazy
so like, objects may need to access the engine later, and if they do, its going to be a bitch to add that. So im adding the engine into the research delegates, just in case. May turn out to be a non-necessity, but i can imagine cases where it would happen, assuming the design doesn't change later.